Implement useMap hook
