+++ project/src/components/map/map.jsx
@@ -1,6 +1,7 @@
-import React, { useRef, useEffect, useState } from 'react';
+import React, { useRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import locationProp from '../../prop-types/location.prop';
+import useMap from '../../hooks/useMap';
 import leaflet from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 
@@ -20,34 +21,8 @@ const defaultIcon = leaflet.icon(defaultMarkerConfig);
 const activeIcon = leaflet.icon(activeMarkerConfig);
 
 function Map({ city, points, selectedPoint }) {
-  const [ map, setMap ] = useState(null);
   const mapRef = useRef(null);
-
-  useEffect(() => {
-    if (mapRef.current !== null && map === null) {
-
-      const mapInstance = leaflet.map(mapRef.current, {
-        center: {
-          lat: city.location.latitude,
-          lng: city.location.longitude,
-        },
-        zoom: city.location.zoom,
-        zoomControl: false,
-        marker: true,
-      });
-
-      leaflet
-        .tileLayer(
-          'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png',
-          {
-            attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>',
-          },
-        )
-        .addTo(mapInstance);
-
-      setMap(mapInstance);
-    }
-  }, [ mapRef, city, map ]);
+  const map = useMap(mapRef, city.location);
 
   useEffect(() => {
     if (map) {
